requirements.txt
asgiref==3.8.1
Django==5.2
django-redis==5.4.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pika==1.3.2
pillow==11.1.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
sqlparse==0.5.3
uritemplate==4.1.1
